requirements.txt
nuna-sql-tools==0.0.30
dataclass-csv==1.4.0
pytz==2022.5
pyarrow==7.0.0
smart_open==6.0.0
pyspark==3.2.1
Faker==8.13.2
absl-py==0.13.0
numpy==1.21.2
pandas==1.3.3
